refactor(optimize): split single-objective module by category
single.py (840 lines, 18 objective classes) is split into focused category
modules: return_objectives.py, risk_objectives.py, and trade_quality_objectives.py
(matching the existing test categorization). single.py re-exports all 18 for
backward compatibility, so the ~8 importers stay stable. 145 optimize tests pass.
